requirements.txt
python-telegram-bot==13.11
requests
python-dotenv
sqlalchemy==2